fix: give running application a chance to stop gracefully
When `hal` asks the running application to stop, it does so by asking
`supervisord` to `ctl stop run`. The `supervisord` configuration doesn't
have a `stopsignal` configuration, so `supervisord` proceeds to directly
sending `SIGKILL`. It's a good practice that applications react to `SIGTERM`
and perform graceful shutdown. This commit therefore configures `supervisord`
to first send `SIGTERM` and only if the application is still running after
5 seconds, proceed with `SIGKILL`.